+++ eslint.config.mjs
@@ -1,41 +1,31 @@
-import { defineConfig } from "eslint/config";
-import globals from "globals";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
+import globals from 'globals'
+import js from '@eslint/js'
+import { defineConfig } from 'eslint/config'
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-    baseDirectory: __dirname,
-    recommendedConfig: js.configs.recommended,
-    allConfig: js.configs.all
-});
-
-export default defineConfig([{
-    extends: compat.extends(
-        "eslint:recommended",
-        "plugin:promise/recommended",
-        "plugin:node/recommended",
-    ),
-
+export default defineConfig([
+  {
     languageOptions: {
-        globals: {
-            ...globals.node,
-            ...globals.mocha,
-        },
-
-        ecmaVersion: 5,
-        sourceType: "module",
+      ecmaVersion: 'latest',
+      globals: {
+        ...globals.node,
+        ...globals.browser,
+        ...globals.mocha,
+      },
+      sourceType: 'module',
     },
 
     rules: {
-        "no-unused-vars": ["error", {
-            args: "none",
-        }],
-
-        "dot-notation": "error",
-        "prefer-const": "warn",
+      // 'no-undef': [ 'warn' ],
+      'no-unused-vars': [
+        'warn',
+        {
+          args: 'none',
+          destructuredArrayIgnorePattern: '^_',
+        },
+      ],
+      'dot-notation': 'error',
+      'prefer-const': 'warn',
     },
-}]);
+  },
+  js.configs.recommended,
+])