+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 #### Changed
 
+- dep: changed test runner from mocha to node:test
 - deps(all): bump to latest
   - dep(eslint): updated to v10
 - style: added prettier config